+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Who knows what the future holds...
 Games:
 - [Don't Starve Together](https://store.steampowered.com/app/322330/Dont_Starve_Together/) support.
 - [Colony Survival](https://store.steampowered.com/app/366090/Colony_Survival/) support.
+- [Onset](https://store.steampowered.com/app/1105810/Onset/) support.
 
 ### Breaking:
 Nothing.